Get rid of some -Wconversion warnings.
git-svn-id: https://svn.r-project.org/R/trunk@55417 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/unix/Rscript.c
@@ -116,7 +116,7 @@ int main(int argc, char *argv[])
 	usage();
 	exit(1);
     }
-    av = (char **) malloc((argc+4)*sizeof(char *));
+    av = (char **) malloc((size_t) (argc+4)*sizeof(char *));
     if(!av) {
 	fprintf(stderr, "malloc failure\n");
 	exit(1);
